Refactored a bunch of class namespaces Hgm -> Hgm.Game
