+++ Hedgemen/Game/EntityComponents/PartyMember.cs
@@ -1,6 +1,6 @@
 using Petal.Framework.EC;
 
-namespace Hgm.EntityComponents;
+namespace Hgm.Game.EntityComponents;
 
 public sealed class PartyMember : EntityComponent
 {
