+++ Hedgemen/Game/Systems/MovementSystem.cs
@@ -1,4 +1,4 @@
-namespace Hgm.Systems;
+namespace Hgm.Game.Systems;
 
 public static class MovementSystem // stub
 {
