+++ Hedgemen/Hedgemen.csproj
@@ -30,11 +30,6 @@
             <HintPath>..\Petal.Framework\lib\FNA.dll</HintPath>
         </Reference>
     </ItemGroup>
-
-    <ItemGroup>
-      <Folder Include="Manifests\" />
-      <Folder Include="WorldGeneration\New\" />
-    </ItemGroup>
 </Project>
 
 <!--i'm bad at remembering things-->
