+++ Hedgemen/Vanilla/HedgemenContent.cs
@@ -2,9 +2,10 @@ using System;
 using System.IO;
 using System.Text;
 using System.Text.Json;
+using Hgm.Game;
 using Hgm.Components;
 using Hgm.Vanilla.WorldGeneration;
-using Hgm.WorldGeneration;
+using Hgm.Game.WorldGeneration;
 using Microsoft.Xna.Framework;
 using Petal.Framework;
 using Petal.Framework.Assets;
