+++ Hedgemen/Vanilla/HedgemenDebugMod.cs
@@ -6,9 +6,10 @@ using System.IO;
 using System.Text;
 using System.Threading.Tasks;
 using Hgm.Components;
-using Hgm.EntityComponents;
+using Hgm.Game.EntityComponents;
+using Hgm.Game;
 using Hgm.Vanilla.WorldGeneration;
-using Hgm.WorldGeneration;
+using Hgm.Game.WorldGeneration;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Petal.Framework;
